move getComponentName to folder, modify readme structure
