+++ README.md
@@ -11,20 +11,4 @@ npm install --save nymag-handlebars
 
 # Utils
 
-### getComponentName
-
-Get component name from uri
-
-#### Params
-
-* `uri` _string_
-
-**Returns** _string|null_
-
-#### Example
-
-```js
-getComponentName('nymag.com/press/components/base/instances/foobarbaz@published')
-//=> 'base'
-
-```
+- getComponentName [code|tests][https://github.com/nymag/clay-utils/tree/develop/lib/getComponentName]
