+++ index.js
@@ -1,3 +1,3 @@
 'use strict';
 
-module.exports.getComponentName = require('./lib/getComponentName');
+module.exports.getComponentName = require('./lib/getComponentName/getComponentName');
