+++ lib/getComponentName/README.md
@@ -1,0 +1,17 @@
+### getComponentName
+
+Get component name from uri
+
+#### Params
+
+* `uri` _string_
+
+**Returns** _string|null_
+
+#### Example
+
+```js
+getComponentName('nymag.com/press/components/base/instances/foobarbaz@published')
+//=> 'base'
+
+```
