+++ lib/getComponentName/getComponentName.js
