+++ lib/getComponentName/getComponentName.test.js
